+++ src/ts/main.tsx
@@ -157,7 +157,7 @@ import '@/features/EWT';
 import '@/features/Einstellungen';
 import './core/orchestration/auth';
 import '@/features/Neben';
-import '@/features/EA';
+import '@/app/features';
 
 initializeAppBootstrap();